- Isolate OSLogin service account derivation from google metadata server to  OSLogin step only
